File 3/backend/emosense/requirements.txt
absl-py==2.1.0
asgiref==3.8.1
astunparse==1.6.3
Brotli==1.1.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.5
defusedxml==0.7.1
distlib==0.3.8
Django==5.0.4
django-allauth==0.61.1
django-cors-headers==4.3.1
djangorestframework==3.15.1
dnspython==2.5.0
filelock==3.13.4
flatbuffers==24.3.25
gast==0.5.4
google-pasta==0.2.0
grpcio==1.62.1
h5py==3.11.0
idna==3.7
keras==3.2.1
libclang==18.1.1
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
ml-dtypes==0.3.2
mysqlclient==2.2.4
namex==0.0.8
numpy==1.26.4
oauthlib==3.2.2
opencv-python==4.9.0.80
opt-einsum==3.3.0
optree==0.11.0
packaging==24.0
pillow==10.3.0
platformdirs==4.2.0
protobuf==4.25.3
pycparser==2.22
Pygments==2.17.2
PyJWT==2.8.0
pymongo==4.6.1
python3-openid==3.2.0
requests==2.31.0
requests-oauthlib==2.0.0
rich==13.7.1
six==1.16.0
sqlparse==0.5.0
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow
tensorflow-intel
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.4.0
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
virtualenv==20.25.3
Werkzeug==3.0.2
whitenoise==6.6.0
wrapt==1.16.0
psycopg2
psycopg